rename: Point, Date 관련 코드 common 패키지 이동
