+++ src/main/java/com/babyblackdog/ddogdog/reservation/service/ReservationFacade.java
@@ -1,9 +1,9 @@
 package com.babyblackdog.ddogdog.reservation.service;
 
-import com.babyblackdog.ddogdog.common.Point;
+import com.babyblackdog.ddogdog.common.date.StayPeriod;
+import com.babyblackdog.ddogdog.common.point.Point;
 import com.babyblackdog.ddogdog.place.reader.PlaceReaderService;
 import com.babyblackdog.ddogdog.place.reader.vo.RoomSimpleResult;
-import com.babyblackdog.ddogdog.reservation.service.dto.StayPeriod;
 import com.babyblackdog.ddogdog.reservation.service.dto.result.OrderedReservationResult;
 import com.babyblackdog.ddogdog.reservation.service.dto.result.RoomOrderPageResult;
 import com.babyblackdog.ddogdog.user.service.UserService;
